Fix failure detection, jsxhint bug, and improve style (#13)
* Fixing bug with jsxhint usage where it only matches the first testcase>temp file name

* Updating package version to 0.2.2 to reflect fix to jsxhint reporting

* Improve style

* Fix failure detection

Failures for the current file are output if the file's name
exists in the `files` object. This object is populated with filenames by
stripping dot directories from the front of the path given by jshint.
Unfortunately, only two characters were being stripped: './'. This is
insufficient on Windows and for files outside of the reporterOutput(?)
directory. Thus, the regex was updated to strip any number of '.', '\',
or '/' characters from the start of the file's path.

The version was updated to reflect this bug fix.